fix(runtime): dedupe string-mode SSR stylesheets via shared hasStylesheetLink
String-mode SSR (LoadableCollector.emitStyleAssets) deduped injected route
stylesheets against every <link href>, so a <link rel="prefetch"> for the same
css URL (e.g. from performance.prefetch) skipped the real <link rel="stylesheet">
and the route rendered unstyled. Extract the stylesheet-aware hasStylesheetLink
helper (added for streaming SSR in #8694) into core/server/utils and reuse it in
both the string and streaming paths.
